mobile-shopping-agent-v2/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
google-generativeai==0.3.1
pydantic==2.4.2
python-multipart==0.0.6
supabase==1.0.3